chore: refactor storybook package chore: fix versions
