+++ .prettierignore
@@ -16,6 +16,7 @@ yarn.lock
 *.yml
 *.html
 *.sh
+*.log
 packages
 settings.gradle
 build.gradle
