+++ .storybook.native/index.js
@@ -5,13 +5,13 @@ import {
   configure,
   addDecorator
 } from "@storybook/react-native";
+import { BarSpacingDecorator } from "@times-components/storybook/decorators";
+import { WhiteBgColorDecorator } from "@times-components/storybook/decorators";
 import { loadStories } from "./story-loader";
 import "./addons";
-import IosBarSpacingDecorator from "../storybook/decorators/ios-bar-spacing";
-import WhiteBgColorDecorator from "../storybook/decorators/white-bg-color";
 
 if (Platform.OS === "ios") {
-  addDecorator(IosBarSpacingDecorator);
+  addDecorator(BarSpacingDecorator);
 }
 
 if (Platform.OS === "android") {
