+++ package.json
@@ -93,9 +93,9 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/addon-actions": "3.2.11",
-    "@storybook/react": "3.2.11",
-    "@storybook/react-native": "3.2.11",
+    "@storybook/addon-actions": "3.2.17",
+    "@storybook/react": "3.2.17",
+    "@storybook/react-native": "3.2.17",
     "@times-components/fructose": "3.1.8",
     "chromeless": "1.3.0",
     "coveralls": "2.13.1",
@@ -113,6 +113,7 @@
     "lerna": "2.5.1",
     "node-fetch": "1.7.2",
     "prettier": "1.8.2",
+    "prop-types": "15.6.0",
     "react": "16.0.0",
     "react-dom": "16.0.0",
     "react-native": "0.50.1",
