+++ packages/ad/package.json
@@ -43,7 +43,7 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/react-native": "3.2.11",
+    "@storybook/react-native": "3.2.17",
     "@times-components/eslint-config-thetimes": "0.0.1",
     "babel-cli": "6.24.1",
     "babel-core": "6.24.1",
