+++ packages/article-image/package.json
@@ -42,9 +42,9 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/react-native": "3.2.11",
+    "@storybook/react-native": "3.2.17",
     "@times-components/eslint-config-thetimes": "0.0.1",
-    "@times-components/utils": "0.0.2",
+    "@times-components/utils": "0.0.5",
     "babel-cli": "6.24.1",
     "babel-core": "6.24.1",
     "babel-jest": "21.2.0",
