+++ packages/article/package.json
@@ -39,7 +39,7 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/react-native": "3.2.11",
+    "@storybook/react-native": "3.2.17",
     "@times-components/eslint-config-thetimes": "0.0.1",
     "@times-components/jest-configurator": "0.0.11",
     "@times-components/utils": "0.0.5",
