+++ packages/author-profile/package.json
@@ -33,9 +33,10 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/addon-actions": "3.2.11",
-    "@storybook/react-native": "3.2.11",
+    "@storybook/addon-actions": "3.2.17",
+    "@storybook/react-native": "3.2.17",
     "@times-components/eslint-config-thetimes": "0.0.1",
+    "@times-components/tealium": "0.0.1",
     "@times-components/utils": "0.0.5",
     "apollo-utilities": "1.0.2",
     "babel-cli": "6.24.1",
