+++ packages/error-view/package.json
@@ -42,8 +42,8 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/addon-actions": "3.2.11",
-    "@storybook/react-native": "3.2.11",
+    "@storybook/addon-actions": "3.2.17",
+    "@storybook/react-native": "3.2.17",
     "@times-components/brightcove-video": "1.8.4",
     "@times-components/eslint-config-thetimes": "0.0.1",
     "babel-cli": "6.24.1",
