+++ packages/link/link.stories.js
@@ -2,8 +2,8 @@ import { View } from "react-native";
 import React from "react";
 import { storiesOf } from "dextrose/storiesOfOverloader";
 import { action } from "@storybook/addon-actions";
+import { CenteredDecorator } from "@times-components/storybook/decorators";
 import Link, { TextLink } from "./link";
-import CenteredDecorator from "../../storybook/decorators/centered";
 
 const BigPinkSquare = props => (
   <View
