+++ packages/pagination/package.json
@@ -32,9 +32,10 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/addon-actions": "3.2.11",
-    "@storybook/react-native": "3.2.11",
+    "@storybook/addon-actions": "3.2.17",
+    "@storybook/react-native": "3.2.17",
     "@times-components/eslint-config-thetimes": "0.0.1",
+    "@times-components/storybook": "0.0.1",
     "babel-cli": "6.24.1",
     "babel-core": "6.24.1",
     "babel-jest": "21.2.0",
