+++ packages/provider/package.json
@@ -41,7 +41,7 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/react-native": "3.2.11",
+    "@storybook/react-native": "3.2.17",
     "@times-components/eslint-config-thetimes": "0.0.1",
     "@times-components/utils": "0.0.5",
     "apollo-utilities": "1.0.2",
