+++ packages/responsive-styles/package.json
@@ -29,9 +29,9 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/react-native": "3.2.11",
+    "@storybook/react-native": "3.2.17",
     "@times-components/eslint-config-thetimes": "0.0.1",
-    "@times-components/jest-configurator": "0.0.8",
+    "@times-components/jest-configurator": "0.0.11",
     "babel-cli": "6.24.1",
     "babel-core": "6.24.1",
     "babel-jest": "21.2.0",
