+++ packages/storybook/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@times-components/storybook",
+  "version": "0.0.1",
+  "description": "React storybook helpers for Times Components",
+  "scripts": {
+    "fmt": "prettier --write '**/*.*'",
+    "prettier:diff": "prettier --list-different '**/*.*'",
+    "lint": "eslint . && yarn prettier:diff"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/newsuk/times-components.git"
+  },
+  "license": "BSD-3-Clause",
+  "bugs": {
+    "url": "https://github.com/newsuk/times-components/issues"
+  },
+  "homepage": "https://github.com/newsuk/times-components#readme",
+  "devDependencies": {
+    "prettier": "1.8.2"
+  },
+  "dependencies": {
+    "@storybook/addon-actions": "3.2.17",
+    "eslint": "4.9.0",
+    "eslint-config-airbnb": "16.1.0",
+    "eslint-config-prettier": "2.6.0",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-jsx-a11y": "6.0.2",
+    "eslint-plugin-react": "7.4.0",
+    "react": "16.0.0",
+    "react-dom": "16.0.0",
+    "react-native": "0.50.1"
+  }
+}
