+++ packages/tealium/.eslintrc.json
@@ -1,0 +1,28 @@
+{
+  "extends": ["airbnb", "prettier", "prettier/react"],
+  "rules": {
+    "react/jsx-filename-extension": [1, { "extensions": [".js"] }],
+    "import/extensions": "off",
+    "import/no-extraneous-dependencies": [
+      "error",
+      {
+        "devDependencies": [
+          "**/__tests__/**",
+          "**/*.test.*",
+          "**/fixtures/**",
+          "**/*.stories*",
+          "**/scripts/**"
+        ],
+        "optionalDependencies": false
+      }
+    ]
+  },
+  "settings": {
+    "import/resolver": {
+      "node": {
+        "extensions": [".js", ".android.js", ".ios.js", ".web.js"]
+      }
+    }
+  },
+  "root": true
+}
