+++ packages/tealium/.npmignore
@@ -1,0 +1,8 @@
+*.stories.js
+__tests__
+*.fructose.js
+*.test.js
+fixtures
+coverage
+android/**/test
+android/**/androidTest
