+++ packages/tealium/__tests__/tealium-send-scheduler.web.test.js
@@ -1,4 +1,6 @@
-import TealiumSendScheduler from "../tealium/tealium-send-scheduler";
+/* eslint-env jest */
+
+import TealiumSendScheduler from "../tealium-send-scheduler";
 
 describe("TealiumSendScheduler", () => {
   const trackingOptions = {
