+++ packages/tracking/tracking.js
@@ -1,4 +1,3 @@
 export { default as withTrackingContext } from "./tracking-context";
 export { default as withTrackEvents } from "./track-events";
 export { default as withTrackScrollDepth } from "./track-scroll-depth";
-export { default as tealiumReporter } from "./tealium/tealium-reporter";
